opensearch/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
motor==3.4.0
multidict==6.0.5
opensearch-py==2.4.2
orjson==3.9.15
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.6.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.6.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
yarl==1.9.4
